requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-amazon
boto3
beautifulsoup4
selenium
requests
watchtower
python-dotenv